Forward Chaining & Other Changes
Forward chaining now works.

Inference Engine is updated so that all literals refer to the same
object in memory. This means when comparing literals you can write this:
literal1 == literal2
rather than:
literal1.getName().equals( liteal2.getName() )

In Method.java
The solve method is renamed to isSolvable to reflect what it's actually
doing. At the end of this method, if solutions are found, set
fReadyToSolve to true.